requirements.txt
pip==23.0.1
Django==5.2.1
asgiref==3.8.1
sqlparse==0.5.3
typing_extensions==4.13.2
setuptools==65.5.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
django-filter==25.1
markdown==3.8
djangorestframework==3.16.0
django-ckeditor-5==0.2.18